requirements.txt
# Kommunales Wissensmanagement mit RAG
# Self-hosted RAG system for German public administration

streamlit>=1.30.0
pypdf2>=3.0.0
python-docx>=1.0.0
chromadb>=0.4.22
requests>=2.31.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
sentence-transformers>=2.2.0
